timeout fix

+++ server.php
@@ -86,9 +86,12 @@ if (isset($_GET['serverip']) && isset($_GET['task'])){
 			break;
 		case 'elastic':
 			$curTime = microtime(true);
-			$stream = ssh2_exec($connection, "curl -s -o /dev/null -XGET http://`/sbin/ifconfig eth1 | grep 'inet addr:' | cut -d: -f2 | awk '{ print $1}'`:9200/_cluster/health?pretty");
-			$timeConsumed = round(microtime(true) - $curTime,3)*1000; 
-			echo "<b><font color='black'>" .$timeConsumed. " ms</font></b>";
+			if (ssh2_exec($connection, "curl -s -o /dev/null -XGET http://`/sbin/ifconfig eth1 | grep 'inet addr:' | cut -d: -f2 | awk '{ print $1}'`:9200/_cluster/health?pretty")) {
+				$timeConsumed = round(microtime(true) - $curTime,3)*1000; 
+				echo "<b><font color='black'>" .$timeConsumed. " ms</font></b>";
+			} else {
+				echo "<b><font color='red'>Timeout</font></b>";
+			}
 			break;
 		default:
 			echo "Wrong task";
